// ProcMgt/ClientApp/src/app/models/designationbusinessunit.ts
import { Businessunits } from './businessunits';

export class Designationbusinessunit {
  
  designationBusinessUnitID: string;
  designationID: string;
  businessUnitTypeID: string;
  businessUnitsID: string;
  userId: string;
  DesignationName: string;
  BusinessUnitTypeName: string;
  BusinessUnitsName: string;

  businessUnits: Businessunits[];

}